+++ Nos/Views/Components/NosNavigationStack.swift
@@ -1,13 +1,5 @@
 import SwiftUI
 
-/// An enumeration of the views that can be pushed onto a `NosNavigationStack`.
-enum NosNavigationDestination: Hashable {
-    case note(RawEventID?)
-    case author(RawAuthorID?)
-    case url(URL)
-    case replyTo(RawEventID?)
-}
-
 /// A `NavigationStack` that knows how to present views common to all the tabs like `Events` and `Authors`.
 /// Take care not to nest these.
 struct NosNavigationStack<Content: View>: View {
